Added new gulp serve task
