+++ client/common/account/authService.js
@@ -1,8 +1,8 @@
 (function(app) {
-  
+
   var authService = function($http, $q, identityService, userService) {
     var authFactory = {};
-    
+
     authFactory.authenticateUser = function(username, password) {
       var deferred = $q.defer();
       $http.post('/api/auth/login', {username:username, password:password}).then(function(response) {
@@ -15,9 +15,9 @@
           deferred.resolve(false);
         }
       });
-      return deferred.promise; 
+      return deferred.promise;
     };
-    
+
     authFactory.logoutUser = function() {
       var deferred = $q.defer();
       $http.post('/api/auth/logout', { logout: true }).then(function() {
@@ -26,7 +26,7 @@
       });
       return deferred.promise;
     };
-    
+
     authFactory.routeAuth = function(role) {
       if(identityService.isAuthorized(role)) {
         return true;
@@ -34,11 +34,11 @@
         return $q.reject('unauthorized');
       }
     }
-    
+
     return authFactory;
   };
-   
-  
+
+
   app.factory('authService', ['$http', '$q', 'identityService', 'userService', authService]);
-  
-}(angular.module('yadaApp')));
+
+}(angular.module('commonUtil')));
