+++ client/common/account/identityService.js
@@ -1,36 +1,36 @@
 (function(app) {
-  
+
   var identityService = function($http, $q, userService) {
     var identityFactory = {};
-                           
+
     identityFactory.currentUser = undefined;
-    
+
     identityFactory.getCurrentUser = function() {
-      var deferred = $q.defer();      
+      var deferred = $q.defer();
       $http.get('/api/auth/currentUser').then(function(response) {
-        if(response.data.success) {  
+        if(response.data.success) {
           var user = new userService.UserResource();
-          angular.extend(user, response.data.user);        
+          angular.extend(user, response.data.user);
           identityFactory.currentUser = user;
           deferred.resolve(true);
-        } else {          
+        } else {
           deferred.resolve(false);
         }
       });
       return deferred.promise;
-    };  
-    
+    };
+
     identityFactory.isAuthenticated = function() {
       return !!this.currentUser;
     };
-    
+
     identityFactory.isAuthorized = function(role) {
       return !!this.currentUser && this.currentUser.roles.indexOf(role) > -1;
     };
-    
+
     return identityFactory;
   };
-  
+
   app.factory('identityService', ['$http', '$q', 'userService', identityService]);
-  
-}(angular.module('yadaApp')));
+
+}(angular.module('commonUtil')));
