+++ package.json
@@ -40,6 +40,7 @@
     "mocha": "^2.2.4",
     "node-mongo-seeds": "^2.2.4",
     "prompt": "^0.2.14",
+    "require-dir": "^0.3.0",
     "should": "^6.0.1",
     "sinon": "^1.14.1",
     "supertest": "^1.0.1",
